fix:placeOrder struct field

+++ requests/rest/trade/trade_requests.go
@@ -12,7 +12,7 @@ type (
 		ClOrdID    string            `json:"clOrdId,omitempty"`
 		Tag        string            `json:"tag,omitempty"`
 		ReduceOnly bool              `json:"reduceOnly,omitempty"`
-		Sz         int64             `json:"sz,string"`
+		Sz         float64           `json:"sz,string"`
 		Px         float64           `json:"px,omitempty,string"`
 		TdMode     okex.TradeMode    `json:"tdMode"`
 		Side       okex.OrderSide    `json:"side"`
